actions/langchain_app/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.8.0
dataclasses-json==0.6.7
exceptiongroup==1.2.2
fastavro==1.9.5
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-cohere==0.2.0
langchain-community==0.2.11
langchain-core==0.2.29
langchain-experimental==0.0.64
langchain-text-splitters==0.2.2
langsmith==0.1.98
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
orjson==3.10.7
packaging==24.1
pandas==2.0.3
parameterized==0.9.0
pinecone-client==5.0.0
pinecone-client[grpc]==5.0.0
pinecone-plugin-inference==1.0.2
protobuf>=4.25,<5.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
gcloud
google-api-python-client
pinecone-client[grpc]
python-dotenv
flask